+++ scm-webapp/src/test/java/sonia/scm/api/v2/resources/DiffResultToDiffResultDtoMapperTest.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-    
+
 package sonia.scm.api.v2.resources;
 
 import de.otto.edison.hal.Link;
@@ -35,11 +35,13 @@ import sonia.scm.repository.api.DiffResult;
 import sonia.scm.repository.api.Hunk;
 
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 import java.util.OptionalInt;
 
 import static java.net.URI.create;
+import static java.util.Collections.emptyIterator;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
@@ -60,6 +62,8 @@ class DiffResultToDiffResultDtoMapperTest {
     assertAddedFile(files.get(0), "A.java", "abc", "java");
     assertModifiedFile(files.get(1), "B.ts", "abc", "def", "typescript");
     assertDeletedFile(files.get(2), "C.go", "ghi", "golang");
+    assertRenamedFile(files.get(3), "typo.ts", "okay.ts", "def",  "fixed", "typescript");
+    assertCopiedFile(files.get(4), "good.ts", "better.ts", "def",  "fixed", "typescript");
 
     DiffResultDto.HunkDto hunk = files.get(1).getHunks().get(0);
     assertHunk(hunk, "@@ -3,4 1,2 @@", 1, 2, 3, 4);
@@ -104,7 +108,9 @@ class DiffResultToDiffResultDtoMapperTest {
           deletedLine("c", 3)
         )
       ),
-      deletedFile("C.go", "ghi")
+      deletedFile("C.go", "ghi"),
+      renamedFile("okay.ts", "typo.ts", "fixed", "def"),
+      copiedFile("better.ts", "good.ts", "fixed", "def")
     );
   }
 
@@ -161,6 +167,24 @@ class DiffResultToDiffResultDtoMapperTest {
     assertThat(file.getLanguage()).isEqualTo(language);
   }
 
+  private void assertRenamedFile(DiffResultDto.FileDto file, String oldPath, String newPath, String oldRevision, String newRevision, String language) {
+    assertThat(file.getOldPath()).isEqualTo(oldPath);
+    assertThat(file.getNewPath()).isEqualTo(newPath);
+    assertThat(file.getOldRevision()).isEqualTo(oldRevision);
+    assertThat(file.getNewRevision()).isEqualTo(newRevision);
+    assertThat(file.getType()).isEqualTo("rename");
+    assertThat(file.getLanguage()).isEqualTo(language);
+  }
+
+  private void assertCopiedFile(DiffResultDto.FileDto file, String oldPath, String newPath, String oldRevision, String newRevision, String language) {
+    assertThat(file.getOldPath()).isEqualTo(oldPath);
+    assertThat(file.getNewPath()).isEqualTo(newPath);
+    assertThat(file.getOldRevision()).isEqualTo(oldRevision);
+    assertThat(file.getNewRevision()).isEqualTo(newRevision);
+    assertThat(file.getType()).isEqualTo("copy");
+    assertThat(file.getLanguage()).isEqualTo(language);
+  }
+
   private DiffResult result(DiffFile... files) {
     DiffResult result = mock(DiffResult.class);
     when(result.iterator()).thenReturn(Arrays.asList(files).iterator());
@@ -171,6 +195,7 @@ class DiffResultToDiffResultDtoMapperTest {
     DiffFile file = mock(DiffFile.class);
     when(file.getNewPath()).thenReturn(path);
     when(file.getNewRevision()).thenReturn(revision);
+    when(file.getChangeType()).thenReturn(DiffFile.ChangeType.ADD);
     when(file.iterator()).thenReturn(Arrays.asList(hunks).iterator());
     return file;
   }
@@ -179,6 +204,7 @@ class DiffResultToDiffResultDtoMapperTest {
     DiffFile file = mock(DiffFile.class);
     when(file.getOldPath()).thenReturn(path);
     when(file.getOldRevision()).thenReturn(revision);
+    when(file.getChangeType()).thenReturn(DiffFile.ChangeType.DELETE);
     when(file.iterator()).thenReturn(Arrays.asList(hunks).iterator());
     return file;
   }
@@ -189,7 +215,30 @@ class DiffResultToDiffResultDtoMapperTest {
     when(file.getNewRevision()).thenReturn(newRevision);
     when(file.getOldPath()).thenReturn(path);
     when(file.getOldRevision()).thenReturn(oldRevision);
+    when(file.getChangeType()).thenReturn(DiffFile.ChangeType.MODIFY);
     when(file.iterator()).thenReturn(Arrays.asList(hunks).iterator());
+    return file;
+  }
+
+  private DiffFile renamedFile(String newPath, String oldPath, String newRevision, String oldRevision) {
+    DiffFile file = mock(DiffFile.class);
+    when(file.getNewPath()).thenReturn(newPath);
+    when(file.getNewRevision()).thenReturn(newRevision);
+    when(file.getOldPath()).thenReturn(oldPath);
+    when(file.getOldRevision()).thenReturn(oldRevision);
+    when(file.getChangeType()).thenReturn(DiffFile.ChangeType.RENAME);
+    when(file.iterator()).thenReturn(emptyIterator());
+    return file;
+  }
+
+  private DiffFile copiedFile(String newPath, String oldPath, String newRevision, String oldRevision) {
+    DiffFile file = mock(DiffFile.class);
+    when(file.getNewPath()).thenReturn(newPath);
+    when(file.getNewRevision()).thenReturn(newRevision);
+    when(file.getOldPath()).thenReturn(oldPath);
+    when(file.getOldRevision()).thenReturn(oldRevision);
+    when(file.getChangeType()).thenReturn(DiffFile.ChangeType.COPY);
+    when(file.iterator()).thenReturn(emptyIterator());
     return file;
   }