+++ scm-webapp/src/test/java/sonia/scm/update/repository/CopyMigrationStrategyTest.java
@@ -21,13 +21,13 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-    
+
 package sonia.scm.update.repository;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.junitpioneer.jupiter.TempDirectory;
+import org.junit.jupiter.api.io.TempDir;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import sonia.scm.SCMContextProvider;
@@ -44,7 +44,6 @@ import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-@ExtendWith(TempDirectory.class)
 @ExtendWith(MockitoExtension.class)
 class CopyMigrationStrategyTest {
 
@@ -54,30 +53,30 @@ class CopyMigrationStrategyTest {
   RepositoryLocationResolver locationResolver;
 
   @BeforeEach
-  void mockContextProvider(@TempDirectory.TempDir Path tempDir) {
+  void mockContextProvider(@TempDir Path tempDir) {
     when(contextProvider.getBaseDirectory()).thenReturn(tempDir.toFile());
   }
 
   @BeforeEach
-  void createV1Home(@TempDirectory.TempDir Path tempDir) throws IOException {
+  void createV1Home(@TempDir Path tempDir) throws IOException {
     V1RepositoryFileSystem.createV1Home(tempDir);
   }
 
   @BeforeEach
-  void mockLocationResolver(@TempDirectory.TempDir Path tempDir) {
+  void mockLocationResolver(@TempDir Path tempDir) {
     RepositoryLocationResolver.RepositoryLocationResolverInstance instanceMock = mock(RepositoryLocationResolver.RepositoryLocationResolverInstance.class);
     when(locationResolver.forClass(Path.class)).thenReturn(instanceMock);
     when(instanceMock.createLocation(anyString())).thenAnswer(invocation -> tempDir.resolve((String) invocation.getArgument(0)));
   }
 
   @Test
-  void shouldUseStandardDirectory(@TempDirectory.TempDir Path tempDir) {
+  void shouldUseStandardDirectory(@TempDir Path tempDir) {
     Path target = new CopyMigrationStrategy(contextProvider, locationResolver).migrate("b4f-a9f0-49f7-ad1f-37d3aae1c55f", "some/more/directories/than/one", "git").get();
     assertThat(target).isEqualTo(tempDir.resolve("b4f-a9f0-49f7-ad1f-37d3aae1c55f"));
   }
 
   @Test
-  void shouldCopyDataDirectory(@TempDirectory.TempDir Path tempDir) {
+  void shouldCopyDataDirectory(@TempDir Path tempDir) {
     Path target = new CopyMigrationStrategy(contextProvider, locationResolver).migrate("b4f-a9f0-49f7-ad1f-37d3aae1c55f", "some/more/directories/than/one", "git").get();
     assertThat(target.resolve("data")).exists();
     Path originalDataDir = tempDir